+++ empower-italy-app/src/components/Hero.js
@@ -1,7 +1,7 @@
 function Hero() {
   return (
     <section id="home" className="text-gray-600 body-font">
-      <div className="container mx-auto flex px-5 py-24 md:flex-row flex-col items-center">
+      <div className="container mx-auto flex px-10 py-24 md:flex-row flex-col items-center">
         <div className="lg:max-w-lg lg:w-full md:w-1/2 w-5/6 mb-10 md:mb-0">
           <img className="object-cover object-center rounded" alt="hero" src="../images/hero.png"></img>
         </div>